# apps/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Synapse API Specification
  description: Core HTTP REST API for Synapse, a Discord-scale real-time communication platform.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local API Server
paths:
...
  /guilds/{guildID}/channels:
    get:
      summary: Fetch all guild channels
      security:
        - BearerAuth: []
      parameters:
        - name: guildID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Channel'
    post:
      summary: Create a new channel
      security:
        - BearerAuth: []
      parameters:
        - name: guildID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - type
              properties:
                name:
                  type: string
                type:
                  type: integer
                  description: '0: Text, 1: Voice, 2: Category'
                parent_channel_id:
                  type: string
                topic:
                  type: string
      responses:
        '201':
          description: Channel created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Channel:
      type: object
      properties:
        id:
          type: string
        guild_id:
          type: string
        parent_channel_id:
          type: string
        name:
          type: string
        type:
          type: integer
        position:
          type: integer
        topic:
          type: string
...